+++ pkg/commonconfig/commonconfig.go
@@ -1,6 +1,7 @@
 package commonconfig
 
 import (
+	"context"
 	"fmt"
 	"regexp"
 
@@ -11,6 +12,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/giantswarm/kubectl-gs/v2/internal/key"
+	"github.com/giantswarm/kubectl-gs/v2/pkg/installation"
 	"github.com/giantswarm/kubectl-gs/v2/pkg/scheme"
 )
 
@@ -34,28 +36,47 @@ func (cc *CommonConfig) GetConfigFlags() genericclioptions.RESTClientGetter {
 	return *cc.ConfigFlags
 }
 
-func (cc *CommonConfig) GetProvider() (string, error) {
+func (cc *CommonConfig) GetProviderFromConfig() (string, error) {
+	config, err := cc.GetConfigFlags().ToRESTConfig()
+	if err != nil {
+		return "", microerror.Mask(err)
+	}
+	return cc.extractProviderFromUrl(config.Host), nil
+}
+
+func (cc *CommonConfig) GetProviderFromInstallation(ctx context.Context, athenaUrl string, fallbackToConfig bool) (string, error) {
 	config, err := cc.GetConfigFlags().ToRESTConfig()
 	if err != nil {
 		return "", microerror.Mask(err)
 	}
 
+	i, err := installation.New(ctx, config.Host, athenaUrl)
+	if err == nil {
+		return i.Provider, nil
+	}
+	if fallbackToConfig {
+		return cc.extractProviderFromUrl(config.Host), nil
+	}
+	return "", microerror.Mask(err)
+}
+
+func (cc *CommonConfig) extractProviderFromUrl(url string) string {
 	awsRegexp := regexp.MustCompile(fmt.Sprintf(providerRegexpPattern, key.ProviderAWS))
 	azureRegexp := regexp.MustCompile(fmt.Sprintf(providerRegexpPattern, key.ProviderAzure))
 
 	var provider string
 	switch {
-	case awsRegexp.MatchString(config.Host):
+	case awsRegexp.MatchString(url):
 		provider = key.ProviderAWS
 
-	case azureRegexp.MatchString(config.Host):
+	case azureRegexp.MatchString(url):
 		provider = key.ProviderAzure
 
 	default:
 		provider = key.ProviderOpenStack
 	}
 
-	return provider, nil
+	return provider
 }
 
 func (cc *CommonConfig) GetClient(logger micrologger.Logger) (k8sclient.Interface, error) {